define wrapper for C FAM structs
In C an array of unknown size may appear within a struct definition
as the last member (as long as there is at least one other named member).
This is a known as flexible array member.

Pre C99, the same behavior could be achieved using zero length arrays.

Flexible Array Members are the go-to choice for working with large
amounts of data prefixed by header values.

KVM uses this type of structures (e.g., kvm_cpuid2, kvm_msr_list, etc.)

FamStructWrapper provides a common abstraction for this type of
structures and helps in treating them similarly to Rust Vec.

Signed-off-by: Serban Iorga <seriorga@amazon.com>