tempdir: Refactoring
PathBuf should not be an option as this is not used anywhere.
The code is also simplified now as the path is a regular field.
As a result, as_path now returns Path and not Option<Path>.

Simplified the remove implementation. Refactored Drop such that it
just calls remove function and ignores the result.

Unified imports from std so we don't have each import on one line.

Signed-off-by: Andreea Florescu <fandree@amazon.com>