seek_hole: update documentation
Added high level documentation. Separated existing documentation into
summary and details.

This commit does not include examples for SeekHole trait because we
need a lot of boiler plate code to do that (we need to create a
temporary directory, create a file, write something to the file
and only then we are able to call into seek_data or seek_hole).

Signed-off-by: Andreea Florescu <fandree@amazon.com>