fam: macro to impl FamStruct also takes entries field name
The macro that generates `impl FamStruct` for structs T with flexible
array member now also takes a parameter to be able to specify the
name of the flexible array member.

This is required as not all kernel structs having a flexible array
member use the same field/member name for the respective member.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>