Use target_os=linux for conditional compilation
Currently, this crate uses target_family=unix for
conditional compilation of sections of code that only work
on Linux. This causes problems on other Unix family
operating systems, which fail to compile the Linux-specific
code. Changing the check to target_os=linux correctly
identifies where that code is expected to work, and allows
non-Linux Unixes to use vmm-sys-util and vm-memory.

Signed-off-by: Allison Randal <allison@lohutok.net>