added metric module
The metric module adds astractions for working with metrics.
This module only defines the interface of metrics and it is not intended
for full implementations because each VMM should be able to define their
specific backend.

Merely this module defines an interface that can be used by other
rust-vmm components to define metrics in a generic way.

Signed-off-by: Andreea Florescu <fandree@amazon.com>