sock_ctrl_msg: mark recv_with_fds as unsafe
Writes to arbitrary pointers are unsafe in Rust.  It's the caller's
job to ensure that it's safe for the memory they are writing to can
contain whatever arbitrary bytes are received over the socket.  For
example, it would be unsafe to have an iovec pointing to the return
value of str::as_mut_ptr, because strings can only contain byte
sequences that are valid UTF-8.

Because it's on the caller to make sure they're passing pointers
safely, any function that writes to iovecs has to be marked as unsafe.

Signed-off-by: Alyssa Ross <hi@alyssa.is>