fix(fam): make as_mut_fam_struct() public and unsafe
We previously made as_mut_fam_struct() of FamStructWrapper private so
that we avoid that users modify the length of the header under our feet.
However, this disallows callers to modify other parts of the header
(which is totally safe to do).

Make as_mut_fam_struct() public again, but also mark it as unsafe,
asking users to guarantee that they will not change the length stored
within the header directly, via the mutable reference.

Signed-off-by: Babis Chalios <bchalios@amazon.es>