fam: Remove `PartialEq` trait constraint on `Entry`
Some of the `FAM` structs that exist in the kernel (and which we hence
deal with in `kvm-bindings`) have `Entry`s that are union. But `bindgen`
doesn't derive `PartialEq` for unions, and so trying to actually declare
the FamStructWrapper type for those means we'd need to painfully
handwrite a ton of `PartialEq` impls.

There's no need for `PartialEq` in the trait definition of `FamStruct`,
so let's just drop it and save us a bunch of pain.

Co-authored-by: Patrick Roy <roypat@amazon.co.uk>
Signed-off-by: Ruoqing He <heruoqing@iscas.ac.cn>