fix: return ENOBUFS if CMSG_TRUNC is set always
Currently, if in_fds.len() == 0, but the kernel has some control data
to read, raw_recvmsg() returns Ok((_, 0)), despite CMSG_TRUNC being set.
Return ENOBUFS instead, as would be the case if in_fds.len() > 0 yet
smaller than the number of fds the kernel has for us.

Simply removing the early-return is correct, as the only way to
msg_controllen to be less than sizeof::<cmsghdr>() is when the kernel
has zeroed it out because no control data was available, or because it
was zero all along because in_fds.len() == 0. In both cases,
cmsg_ptr.is_null() will be true, we skip over the loop, and the final
return will end up being Ok((_, 0)).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>