sock_ctrl_msg: Drop unneeded loop
raw_recvmsg implemented a loop to be able to process arbitrary control
message chains. However, in this specific function, we know exactly what
the control message chain looks like, because we construct it ourselves.
It always consists of a single cmsghdr structure. So remove all the code
related to processing different layouts.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>