sock_ctrl_msg: treat cmsghdr as FamStruct
cmsghdr is a fam struct with some additional alignment constraints for
the flexible array member, so we can use the machinery in fam.rs to deal
with it. This comes with the benefit of a _lot_ less unsafe code.

It removes an optimization where sometimes cmsghdr structures were
allocated on the stack if the number of fds to receive was small enough,
but raw_recvmsg was already allocating on the heap for the iovecs
anyway, so this optimization seems a bit pre-mature.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>